team_23-main/machine_learning_models/light_gradient_boosting_machine/requirements.txt
Flask==3.0.3
loguru==0.7.0
lightgbm==4.3.0
protobuf==4.25.5
requests==2.32.3
numpy==1.26.4
pandas==2.2.1
scikit_learn==1.4.0
statsmodels==0.14.1
setuptools>=69.2.0
functions-framework>=3.4.0
google-cloud-bigquery
db-dtypes==1.3.0
matplotlib==3.9.2